refactor: Rename `lib/generateEvent.js`
