+++ lib/generate-event.js
