+++ lib/generate-event.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const { generate } = require('./generateEvent');
+const { generate } = require('./generate-event');
 const zlib = require('zlib');
 
 describe('generating events', () => {
