+++ lib/plugin.js
@@ -15,7 +15,7 @@ const getAppUids = require('./app-uids');
 const removeDestination = require('./removeDestination');
 const { saveDeployment, createAndSetDeploymentUid } = require('./deployment');
 const variables = require('./variables');
-const { generate } = require('./generateEvent');
+const { generate } = require('./generate-event');
 const { configureDeployProfile } = require('./deploy-profile');
 const { test } = require('./test');
 const { getDashboardUrl, dashboardHandler } = require('./dashboard');
