+++ lib/plugin.test.js
@@ -48,7 +48,7 @@ describe('plugin', () => {
       './wrapClean': wrapClean,
       './removeDestination': removeDestination,
       './deployment': { saveDeployment, createAndSetDeploymentUid },
-      './generateEvent': { eventDict: {}, generate },
+      './generate-event': { eventDict: {}, generate },
       './deploy-profile': { configureDeployProfile },
       './injectLogsIamRole': injectLogsIamRole,
       './isAuthenticated': () => true,
